+++ pkg/linkerd-iptables/iptables.go
@@ -12,7 +12,7 @@ import (
 
 	log "github.com/sirupsen/logrus"
 
-	"github.com/linkerd/linkerd2-proxy-init/proxy-init/internal/util"
+	"github.com/linkerd/linkerd2-proxy-init/pkg/linkerd-util"
 )
 
 const (
