+++ proxy-init/cmd/root.go
@@ -7,8 +7,8 @@ import (
 
 	log "github.com/sirupsen/logrus"
 
-	"github.com/linkerd/linkerd2-proxy-init/proxy-init/internal/iptables"
-	"github.com/linkerd/linkerd2-proxy-init/proxy-init/internal/util"
+	"github.com/linkerd/linkerd2-proxy-init/pkg/linkerd-iptables"
+	"github.com/linkerd/linkerd2-proxy-init/pkg/linkerd-util"
 	"github.com/spf13/cobra"
 )
