+++ proxy-init/cmd/root_test.go
@@ -4,7 +4,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/linkerd/linkerd2-proxy-init/proxy-init/internal/iptables"
+	"github.com/linkerd/linkerd2-proxy-init/pkg/linkerd-iptables"
 )
 
 func TestBuildFirewallConfiguration(t *testing.T) {
